Added benchmarking function and saving result function.

+++ include/datapoint.h
@@ -5,12 +5,12 @@
 
 class DataPoint {
     private:
-        double *data; // Pointer to the data array
+        float *data; // Pointer to the data array
         int dim; // Dimension of the data point
         bool visited = false; // Flag to indicate if the point has been visited
     public:
         // Constructor
-        DataPoint(double *values, int dim);
+        DataPoint(float *values, int dim);
 
         // Copy constructor
         DataPoint(const DataPoint& other);
@@ -19,19 +19,19 @@ class DataPoint {
         ~DataPoint();
 
         // Accessor
-        double operator[](int index) const;
+        float operator[](int index) const;
 
         // Copy assignment operator
         DataPoint& operator=(const DataPoint& other);
 
         // Copy method
-        void copy(double *arr, int n) const;
+        void copy(float *arr, int n) const;
 
         // Get Dimension
         int get_dim() const { return dim; }
         
         // Get Data
-        double* get_data() const { return data; }
+        float* get_data() const { return data; }
 };
 
 #endif // DATAPOINT_H_
